rename PrinterSettins folder into Printer folder.
